backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
supabase>=2.22.0
httpx>=0.24.0,<0.29.0
websockets>=15.0.0

# HTTP clients
aiohttp>=3.9.0
httpx>=0.24.0,<0.29.0  # Enhanced for async operations

# AI/LLM
google-generativeai>=0.3.0
openai>=1.3.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
python-dateutil>=2.8.0

# NLP and domain scoring
spacy>=3.7.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File handling
python-multipart>=0.0.6

# PDF generation
reportlab>=4.0.0
weasyprint>=60.0

# Caching
redis>=5.0.0

# Logging
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0